requirements.txt
python-telegram-bot==22.7
httpx>=0.27,<0.29
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
psutil>=5.9.0